% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balassa_index.R
\name{balassa_index}
\alias{balassa_index}
\title{Balassa Index}
\usage{
balassa_index(
  data,
  country = "country",
  product = "product",
  value = "value",
  discrete = TRUE,
  cutoff = 1
)
}
\arguments{
\item{data}{(Type: data.frame, matrix or dgCMatrix) a dataset such as \code{galactic_federation}
or any arrangement.}

\item{country}{(Type: character) the column with the elements of set X.
By default this is set to \code{"country"}.}

\item{product}{(Type: character) the column with the elements of set Y.
By default this is set to \code{"product"}.}

\item{value}{(Type: character) the column with the binary expression for the
Balassa Index.
By default this is set to \code{"value"}.}

\item{discrete}{(Type: logical) whether converting the Balassa Index to
discrete (0/1) values. Anything below the specified cutoff is converted to 0
and 1 otherwise. By default this is set to \code{TRUE}.}

\item{cutoff}{(Type: numeric) the cutoff to use for discretization.
By default this is set to \code{1}.}
}
\value{
A matrix with the Balassa Index.
}
\description{
\code{balassa_index()} computes the Balassa Index for a
bipartite relation between two disjoint sets X, the "country" or "from" side,
and Y, the "product" or "to" side.
}
\details{
The current implementation follows
\insertCite{measuringcomplexity2015}{economiccomplexity} to obtain a metric for
specialisation. In the context of international trade, if the Balassa Index
for a country-product pair is more than 1, it means that country is
specialized in that product.
}
\examples{
balassa_index(
  data = galactic_federation,
  country = "planet",
  product = "product",
  value = "export_value"
)

}
\references{
For more information see:

\insertRef{measuringcomplexity2015}{economiccomplexity}

and the references therein.
}
\keyword{functions}
