% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complexity_outlook.R
\name{complexity_outlook}
\alias{complexity_outlook}
\title{Complexity Outlook}
\usage{
complexity_outlook(balassa_index, proximity_product, complexity_index_product)
}
\arguments{
\item{balassa_index}{(Type: matrix or dgCMatrix) foo}

\item{proximity_product}{(Type: matrix or dgCMatrix) foo.}

\item{complexity_index_product}{(Type: numeric) foobar.}
}
\value{
foobar
}
\description{
foo
}
\details{
bar
}
\examples{
complexity_outlook(
 economiccomplexity_output$balassa_index,
 economiccomplexity_output$proximity$proximity_product,
 economiccomplexity_output$complexity_measures$complexity_index_product
)

}
\references{
For more information on this index see:

\insertRef{atlas2014}{economiccomplexity}

and the references therein.
}
\keyword{functions}
