% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/economiccomplexity-internals.R
\name{country_aggregation}
\alias{country_aggregation}
\title{country aggregation}
\usage{
country_aggregation(dataframe, country = "country", value = "value")
}
\description{
country aggregation
}
\keyword{internal}
