% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/economiccomplexity-internals.R
\name{country_product_aggregation}
\alias{country_product_aggregation}
\title{country-product aggregation}
\usage{
country_product_aggregation(
  dataframe,
  country = "country",
  product = "product",
  value = "value"
)
}
\description{
country-product aggregation
}
\keyword{internal}
