% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{dataframe_to_matrix}
\alias{dataframe_to_matrix}
\title{Dataframe to matrix}
\usage{
dataframe_to_matrix(
  dataframe,
  country = "country",
  product = "product",
  value = "value"
)
}
\description{
Dataframe to matrix
}
\keyword{internal}
