\name{ec_gdp_pc_1962}
\alias{ec_gdp_pc_1962}
\docType{data}
\title{
GDP per capita in 1962
}
\description{
Example dataset with countries and their gross domestic product divided by total population.
}
\usage{ec_gdp_pc_1962}
\format{
  A data frame with 135 observations on the following 2 variables.
  \describe{
    \item{\code{country}}{a character vector}
    \item{\code{value}}{a numeric vector}
  }
}
\source{
The World Bank
}
\examples{
ec_gdp_pc_1962
}
\keyword{datasets}
