\name{ec_output_demo}
\alias{ec_output_demo}
\docType{data}
\title{
Example outputs of the functions within the package
}
\description{
List of different economic complexity computations obtained by using the
functions within the package. The output is provided as a list of data frames.}
\usage{ec_output_demo}
\format{
  A list of data frames with 3 elements.
  \describe{
    \item{\code{rca}}{`rca()` output (single data frame)}
    \item{\code{complexity}}{`complexity()` output (list of four data frames)}
    \item{\code{value}}{`proximity()` output (list of two data frames)}
  }
}
\examples{
ec_output_demo
}
\keyword{datasets}
