% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proximity.R
\name{ec_proximity}
\alias{ec_proximity}
\title{Proximity}
\usage{
ec_proximity(rca, diversity, ubiquity, country_r = "country",
  product_r = "product", value_r = "value", country_d = "country",
  value_d = "value", product_u = "product", value_u = "value",
  tbl = TRUE, compute = "both")
}
\arguments{
\item{rca}{matrix or data.frame with RCA values}

\item{diversity}{matrix or data.frame with diversity values}

\item{ubiquity}{matrix or data.frame with ubiquity values}

\item{country_r}{column containing countries (applies only if d is a
data.frame)}

\item{product_r}{column containing products (applies only if d is a
data.frame)}

\item{value_r}{column containing traded values (applies only if d is a
data.frame)}

\item{country_d}{column containing countries (applies only if d is a
data.frame)}

\item{value_d}{column containing values (applies only if d is a
data.frame)}

\item{product_u}{column containing products (applies only if d is a
data.frame)}

\item{value_u}{column containing values (applies only if d is a
data.frame)}

\item{tbl}{TRUE (default) returns a data.frame and FALSE returns a matrix}

\item{compute}{"country", "product" or "both" (default) matrices}
}
\description{
Proximity
}
\examples{
ec_proximity(
  rca = ec_output_demo$rca_tbl,
  d = ec_output_demo$complexity_measures_tbl$diversity,
  u = ec_output_demo$complexity_measures_tbl$ubiquity
)
}
\references{
For more information on proximity and its applications see:

\insertRef{atlas2014}{economiccomplexity}
}
\keyword{functions}
