\name{ec_trade_1962}
\alias{ec_trade_1962}
\docType{data}
\title{
Exported products in 1962
}
\description{
Example dataset with countries and the products they export. The exports are
grouped to four digits HS 2007 codes.
}
\usage{ec_trade_1962}
\format{
  A data frame with 44435 observations on the following 3 variables.
  \describe{
    \item{\code{country}}{a character vector}
    \item{\code{product}}{a character vector}
    \item{\code{value}}{a numeric vector}
  }
}
\source{
Open Trade Statistics (tradestatistics.io)
}
\examples{
ec_trade_1962
}
\keyword{datasets}
