% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{economic_complexity_measures}
\alias{economic_complexity_measures}
\title{Indices}
\usage{
economic_complexity_measures(d, method = "reflections",
  iterations = 20, output = "matrix")
}
\arguments{
\item{d}{"wide" matrix}

\item{method}{"reflections" or "eigenvalues"}

\item{iterations}{only used for the reflections method (default set to 20)}

\item{output}{"matrix" or "tibble"}
}
\description{
Indices
}
\examples{
rca <- rca(d = world_trade_1980, c = "reporter_iso",
    p = "product_code", x = "export_value_usd")

indices <- indices(rca, method = "reflections", iterations = 20,
    output = "matrix")
}
\keyword{functions}
