% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/economiccomplexity-package.R
\docType{package}
\name{economiccomplexity-package}
\alias{economiccomplexity}
\alias{economiccomplexity-package}
\title{economiccomplexity: Computational Methods for Economic Complexity}
\description{
A wrapper of different methods from Linear Algebra for the equations introduced in The Atlas of Economic Complexity and related literature. This package provides standard matrix and graph output that can be used seamlessly with other packages. See \doi{10.21105/joss.01866} for a summary of these methods and its evolution in literature.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://pacha.dev/economiccomplexity/}
  \item Report bugs at \url{https://github.com/pachadotdev/economiccomplexity/issues/}
}

}
\author{
\strong{Maintainer}: Mauricio Vargas Sepulveda \email{m.sepulveda@mail.utoronto.ca} (\href{https://orcid.org/0000-0003-1017-7574}{ORCID}) [copyright holder]

Other contributors:
\itemize{
  \item Carlo Bottai (improved the eigenvalues calculation) [contributor]
  \item Diego Kozlowski (provided initial RCA function used up to v0.2.4) [contributor]
  \item Nico Pintar (suggested improvements to special cases in eigenvalues calculation) [reviewer]
  \item  The World Bank (World GDP per capita dataset) [data contributor]
  \item  Open Trade Statistics (World Trade dataset) [data contributor]
}

}
\keyword{internal}
