\name{economiccomplexity_output}
\alias{economiccomplexity_output}
\docType{data}
\title{
Example outputs of the functions within the package
}
\description{
List of different economic complexity computations obtained by using the functions within the
package. The output is provided as a list of tibbles.}
\usage{economiccomplexity_output}
\format{
  A list of sparse matrices.
}
\examples{
economiccomplexity_output$balassa_index
economiccomplexity_output$complexity_measures$complexity_index_x
}
\keyword{datasets}
