\name{galactic_federation}
\alias{galactic_federation}
\docType{data}
\title{
Fictional data based on Rick and Morty
}
\description{
A matrix with fictional trade for the planets that were part of
the Galatic Federation in 2017 according to Rick and Morty. The planets are
taken from the TV show but the products and values are just an invention. Each
row expresses how many Blemflarcks (Galatic Federation currency) each planet
exported to the rest of the interplanetary trade system.
}
\usage{galactic_federation}
\format{
  A sparse matrix with 9 rows and 12 columns.
}
\examples{
galactic_federation
}
\keyword{datasets}
