\name{gdp_and_population_2017}
\alias{gdp_and_population_2017}
\docType{data}
\title{
gdp_and_population_2017
}
\description{
tbd
}
\usage{data("gdp_and_population_2017")}
\format{
  A data frame with 262 observations on the following 4 variables.
  \describe{
    \item{\code{reporter_iso}}{a character vector}
    \item{\code{gdp_current_usd}}{a numeric vector}
    \item{\code{population_total}}{a numeric vector}
    \item{\code{gdp_pc_usd}}{a numeric vector}
  }
}
\examples{
data(gdp_and_population_2017)
## maybe str(gdp_and_population_2017) ; plot(gdp_and_population_2017) ...
}
\keyword{datasets}
