% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{indices}
\alias{indices}
\title{RCA (Revealed Comparative Advantage)}
\usage{
indices(m, maxiter = 20, method = "reflections", output = "matrix")
}
\arguments{
\item{m}{"wide" matrix}

\item{maxiter}{20}

\item{method}{"reflections" or "eigenvalues"}

\item{output}{"matrix" or "tibble"}
}
\description{
RCA (Revealed Comparative Advantage)
}
\examples{
rca <- rca(d = world_trade_1980, c = "reporter_iso",
    p = "product_code", x = "export_value_usd")

indices <- indices(rca, method = "reflections", maxiter = 20,
    output = "matrix")
}
\keyword{functions}
