% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prody.R
\name{prody}
\alias{prody}
\title{Revealed Comparative Advantage (RCA)}
\usage{
prody(d1 = NULL, c1 = NULL, p1 = NULL, v1 = NULL, d2 = NULL,
  c2 = NULL, v2 = NULL, tbl_output = FALSE)
}
\arguments{
\item{d1}{tibble/data.frame in long format, it must contain the columns country (character/factor),
product (character/factor) and export value (numeric)}

\item{c1}{string to indicate the column that contains exporting countries in d1 (e.g. "reporter_iso")}

\item{p1}{string to indicate the column that contains exported products in d1 (e.g. "product_code")}

\item{v1}{string to indicate the column that contains traded values in d1(e.g. "trade_value_usd")}

\item{d2}{tibble/data.frame in long format, it must contain the columns country (character/factor),
and GDP per capita (numeric)}

\item{c2}{string to indicate the column that contains exporting countries in d2 (e.g. "reporter_iso")}

\item{v2}{string to indicate the column that contains GDP per capita in d2 (e.g. "gdp_pc")}

\item{tbl_output}{when set to TRUE the output will be a tibble instead of a matrix (default set to FALSE)}
}
\description{
Revealed Comparative Advantage (RCA)
}
\examples{
prody <- prody(
    d1 = world_trade_2017, c1 = "reporter_iso", p1 = "product_code", v1 = "export_value_usd",
    d2 = gdp_and_population_2017, c2 = "reporter_iso", v2 = "gdp_pc_usd"
)
}
\keyword{functions}
