% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projections.R
\name{projections}
\alias{projections}
\title{Projections of a bipartite network}
\usage{
projections(proximity_source, proximity_target, avg_links = 4,
  tolerance = 0.05, compute = "both")
}
\arguments{
\item{proximity_source}{a data frame containing proximity
values for the elements of set X (e.g. proximity_source from \code{proximity()})}

\item{proximity_target}{a data frame containing proximity
values for the elements of set Y (e.g. proximity_target from \code{proximity()})}

\item{avg_links}{average number of connections for the projection of X
(default set to 4)}

\item{tolerance}{tolerance for proximity variation on each iteration until
obtaining the desired average number of connections (default set to 0.05)}

\item{compute}{which projection to compute. By default is "both" (both
projections) but it can also be "source" or "target".}
}
\description{
TBD
}
\details{
TBD
}
\examples{
projections(
  proximity_source = economiccomplexity_output$proximity$proximity_source,
  proximity_target = economiccomplexity_output$proximity$proximity_target
)
}
\references{
For more information see:

\insertRef{atlas2014}{economiccomplexity}

and the references therein.
}
\keyword{functions}
