% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proximity.R
\name{proximity}
\alias{proximity}
\title{Proximity}
\usage{
proximity(
  balassa_index,
  balassa_sum_country,
  balassa_sum_product,
  compute = "both"
)
}
\arguments{
\item{balassa_index}{a data frame (e.g. the output from
\code{balassa_index()}).}

\item{balassa_sum_country}{numeric vector or tibble/data.frame containing the
Balassa sum for elements of set X (e.g. \code{balassa_sum_country} from
\code{complexity_measures()}).}

\item{balassa_sum_product}{numeric vector or tibble/data.frame containing the
Balassa sum for elements of set Y (e.g. \code{balassa_sum_product} from
\code{complexity_measures()}).}

\item{compute}{which proximity to compute. By default is "both" (both
proximities) but it can also be "country" or "product".}
}
\description{
TBDs
}
\details{
TBD
}
\examples{
proximity(
  balassa_index = economiccomplexity_output$balassa_index,
  balassa_sum_country = economiccomplexity_output$complexity_measures$balassa_sum_country,
  balassa_sum_product = economiccomplexity_output$complexity_measures$balassa_sum_product
)
}
\references{
For more information see:

\insertRef{atlas2014}{economiccomplexity}

and the references therein.
}
\keyword{functions}
