\name{proximity_matrices_2017}
\alias{proximity_matrices_2017}
\docType{data}
\title{
TBD
}
\description{
TBD
}
\usage{data("proximity_matrices_2017")}
\format{
  The format is:
List of 2
 $ countries_proximity:Classes ‘tbl_df’, ‘tbl’ and 'data.frame':	0 obs. of  3 variables:
  ..$ from : chr(0)
  ..$ to   : chr(0)
  ..$ value: num(0)
 $ products_proximity :Classes ‘tbl_df’, ‘tbl’ and 'data.frame':	0 obs. of  3 variables:
  ..$ from : chr(0)
  ..$ to   : chr(0)
  ..$ value: num(0)
}
\details{
TBD
}
\source{
TBD
}
\references{
TBD
}
\examples{
data(proximity_matrices_2017)
## maybe str(proximity_matrices_2017) ; plot(proximity_matrices_2017) ...
}
\keyword{datasets}
