% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proximity_networks.R
\name{proximity_networks}
\alias{proximity_networks}
\title{Networks}
\usage{
proximity_networks(proximity_countries, proximity_products,
  c_cutoff = 0.25, p_cutoff = 0.55, tbl_output = FALSE)
}
\arguments{
\item{proximity_countries}{matrix or tibble/data.frame, if d is a tibble/data.frame it must contain the columns
from (character/factor), to (character/factor) and value (numeric), if it is a matrix it must be a
numeric matrix with countries in row names and column names}

\item{proximity_products}{matrix or tibble/data.frame, if d is a tibble/data.frame it must contain the columns
from (character/factor), to (character/factor) and value (numeric), if it is a matrix it must be a
numeric matrix with products in row names and column names}

\item{c_cutoff}{all the values lower than the specified c_cutoff will be converted to 0 and excluded
from the countries network (default set to 0.2)}

\item{p_cutoff}{all the values lower than the specified p_cutoff will be converted to 0 and excluded
from the products network (default set to 0.4)}

\item{tbl_output}{when set to TRUE the output will be a tibble instead of a graph (default set to FALSE)}
}
\description{
Networks
}
\examples{
networks <- proximity_networks(proximity_matrices_2017$countries_proximity,
    proximity_matrices_2017$products_proximity)
}
\keyword{functions}
