\name{rca_m}
\alias{rca_m}
\docType{data}
\title{
Matrix output of revealed_comparative_advantage()
}
\description{
TBD
}
\usage{data("rca_m")}
\format{
  The format is:
Formal class 'dgCMatrix' [package "Matrix"] with 6 slots
  ..@ i       : int [1:301] 1 8 11 14 15 16 19 23 24 25 ...
  ..@ p       : int [1:12] 0 44 73 119 183 224 247 262 267 281 ...
  ..@ Dim     : int [1:2] 86 11
  ..@ Dimnames:List of 2
  .. ..$ : chr [1:86] "Albania" "Algeria" "Angola" "Anguilla" ...
  .. ..$ : chr [1:11] "1 Transportation" "10 Personal, cultural, and recreational services" "11 Government services, n.i.e." "2 Travel" ...
  ..@ x       : num [1:301] 1 1 1 1 1 1 1 1 1 1 ...
  ..@ factors : list()
}
\examples{
rca_m
}
\keyword{datasets}
