\name{rca_t}
\alias{rca_t}
\docType{data}
\title{
Tibble output of revealed_comparative_advantage()
}
\description{
TBD
}
\usage{data("rca_t")}
\format{
  A data frame with 790 observations on the following 3 variables.
  \describe{
    \item{\code{country}}{a character vector}
    \item{\code{product}}{a character vector}
    \item{\code{value}}{a numeric vector}
  }
}
\examples{
rca_t
}
\keyword{datasets}
