% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complexity_measures.R
\name{reflections_method}
\alias{reflections_method}
\title{Reflections Method}
\usage{
reflections_method(balassa_index, iterations)
}
\arguments{
\item{...}{params inherited from \code{complexity_measures()}}
}
\description{
Reflections Method
}
\keyword{internal}
