\name{rick_and_morty_gdp_numeric}
\alias{rick_and_morty_gdp_numeric}
\docType{data}
\title{
Example dataset with fabricated numbers using Rick and Morty references (numeric version)
}
\description{
This fictional datasets includes planets from Rick and Morty as "countries" and each
"country" has an associated gdp per capita. The currency used here is "brapples".
Please consider this just as a lightweight dataset for examples and tests.
}
\usage{data("rick_and_morty_gdp_numeric")}
\format{
  The format is:
 Named num [1:9] 3800 1800 2600 5000 300 5800 1900 1500 4200
 - attr(*, "names")= chr [1:9] "Alphabetrium" "Bird World" "Earth C-137" "Gazorpazorp" ...
}
\examples{
rick_and_morty_gdp_numeric
}
\keyword{datasets}
