\name{rick_and_morty_trade_tibble}
\alias{rick_and_morty_trade_tibble}
\docType{data}
\title{
Example dataset with fabricated numbers using Rick and Morty references (tibble version)
}
\description{
This fictional datasets includes planets from Rick and Morty as "countries" and each
"country" has an associated gdp per capita. The currency used here is "brapples".
Please consider this just as a lightweight dataset for examples and tests.
}
\usage{data("rick_and_morty_trade_tibble")}
\format{
  A data frame with 9 observations on the following 2 variables.
  \describe{
    \item{\code{country}}{a character vector}
    \item{\code{value}}{a numeric vector}
  }
}
\examples{
rick_and_morty_trade_tibble
}
\keyword{datasets}
