% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/economiccomplexity-internals.R
\name{source_target_aggregation}
\alias{source_target_aggregation}
\title{Source-target aggregation}
\usage{
source_target_aggregation(dataframe, source = "source",
  target = "target", value = "value")
}
\description{
Source-target aggregation
}
\keyword{internal}
