\name{world_gdp_avg_1998_to_2000}
\alias{world_gdp_avg_1998_to_2000}
\docType{data}
\title{
World Trade Per-Capita GDP for the Period 1998-2000
}
\description{
A named vector that summarises the per-capita GDP of different countries.
The unit is year 2000 USD.
}
\usage{world_gdp_avg_1998_to_2000}
\format{
  A named vector with 240 entries.
}
\examples{
world_gdp_avg_1998_to_2000
}
\keyword{datasets}
