\name{world_rca_2017}
\alias{world_rca_2017}
\docType{data}
\title{
TBD
}
\description{
TBD
}
\usage{data("world_rca_2017")}
\format{
  A data frame with 176971 observations on the following 3 variables.
  \describe{
    \item{\code{country}}{a character vector}
    \item{\code{product}}{a character vector}
    \item{\code{value}}{a numeric vector}
  }
}
\details{
TBD
}
\source{
TBD
}
\references{
TBD
}
\examples{
data(world_rca_2017)
## maybe str(world_rca_2017) ; plot(world_rca_2017) ...
}
\keyword{datasets}
