% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/economiccomplexity-package.R
\docType{data}
\name{world_trade_avg_1998_to_2000}
\alias{world_trade_avg_1998_to_2000}
\title{World Trade Averages for the Period 1998-2000}
\format{
A data frame with 124,336 rows and 3 columns.
}
\usage{
data(world_trade_avg_1998_to_2000)
}
\description{
A data frame that summarizes all the products that different countries exported to the rest of the world. This data uses the SITC revision 2 classification with four digits product codes. The unit is year 2000 USD.
}
\examples{
data(world_trade_avg_1998_to_2000)
head(world_trade_avg_1998_to_2000)
}
\keyword{datasets}
