% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complexity.R
\name{complexity}
\alias{complexity}
\title{Complexity measures}
\usage{
complexity(rca, method = "fitness", iterations = 20, extremality = 1,
  atlas = FALSE, tbl = TRUE, country = "country",
  product = "product", value = "value")
}
\arguments{
\item{rca}{matrix or data frame with RCA values}

\item{method}{which method to use (by default is "fitness", it can also
be "reflections" or "eigenvalues")}

\item{iterations}{number of iterations to use in the fitness, reflections
and indirectly in eigenvalues method (by default is 20)}

\item{extremality}{numeric coefficient to use in the fitness method
(by default is 1)}

\item{atlas}{remove the countries not ranked in The Atlas of Economic
Complexity (by default is FALSE)}

\item{tbl}{TRUE (default) returns a data.frame and FALSE returns a matrix}

\item{country}{column containing countries (applies only if d is a
data.frame)}

\item{product}{column containing products (applies only if d is a
data.frame)}

\item{value}{column containing traded values (applies only if d is a
data.frame)}
}
\value{
A list with four data frames or matrices.
}
\description{
\code{rca} computes complexity indices following the definitions
from \insertCite{measuringcomplexity2015;textual}{economiccomplexity}
}
\details{
Given a \eqn{C\times P} matrix (C for "countries"
and P for "products") or an equivalent 3-columns data frame with RCA
values as input, this function implements the equations:
\deqn{\text{(Diversity)}\: k_{c}^{(0)} = \sum_p R_{cp}}
\deqn{\text{(Ubiquity)}\: k_{p}^{(0)} = \sum_c R_{cp}}
\deqn{\text{(Fitness)}\: \tilde{F}^{(n)}(\gamma) = \sum_{p} R_{cp}Q_{p}^{n-1}}
\deqn{\text{(Complexity)}\: \tilde{Q}^{(n)}(\gamma) =
\left[\sum_{c} R_{cp}(F_{c}^{n-1})^{-\gamma}\right]^{-1/\gamma}.}

Besides this implementation, the function offers the reflections method
and eigenvalues computation as alternatives to the Fitness-Complexity
formulation.
}
\examples{
complexity(ec_output_demo$rca)

}
\references{
For more information about complexity measures see:

\insertRef{measuringcomplexity2015}{economiccomplexity}

and the references therein.
}
\seealso{
\code{\link[economiccomplexity]{rca}},
}
\keyword{functions}
