% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complexity_measures.R
\name{complexity_measures}
\alias{complexity_measures}
\title{Complexity Measures}
\usage{
complexity_measures(d = NULL, c = "country", p = "product",
  v = "value", method = "fitness", iterations = 20,
  extremality = 1, keep_atlas = FALSE, tbl_output = FALSE)
}
\arguments{
\item{d}{matrix or tibble/data.frame in long format, if d is a tibble/data.frame it must contain the columns
country (character/factor), product (character/factor) and discrete RCA (integer), if it is a matrix it must be a
zero/one matrix with countries in the row names and products in the column names}

\item{c}{string to indicate the column that contains exporting countries (default set to "country" that is the
output of revealed_comparative_advantage())}

\item{p}{string to indicate the column that contains exported products (default set to "product" that is the
output of revealed_comparative_advantage())}

\item{v}{string to indicate the column that contains RCA values (default set to "value" that is the
output of revealed_comparative_advantage())}

\item{method}{reflections, eigenvalues or fitness (default set to reflections)
methods (default set to FALSE)}

\item{iterations}{number of iterations to use in the reflections method (default set to 20)}

\item{extremality}{coefficient to use in the fitness method (default set to 1)}

\item{keep_atlas}{removes the countries not ranked in The Atlas of Economic Complexity (default set to FALSE)}

\item{tbl_output}{when set to TRUE the output will be a tibble instead of a matrix (default set to FALSE)}
}
\description{
Complexity Measures
}
\examples{
cm <- complexity_measures(world_rca_2017)
}
\references{
For more information on complexity measures, indices and its applications see:

\insertRef{atlas2014}{economiccomplexity}

\insertRef{measuringcomplexity2015}{economiccomplexity}
}
\keyword{functions}
