% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countries_position.R
\name{countries_position}
\alias{countries_position}
\title{Countries position}
\usage{
countries_position(rca = NULL, c1 = "country", p1 = "product",
  v1 = "value", proximity_products = NULL, p21 = "from",
  p22 = "to", v2 = "value", pci = NULL, p3 = "product",
  v3 = "value", tbl_output = FALSE)
}
\arguments{
\item{rca}{matrix or tibble/data.frame in long format (e.g. the output of
revealed_comparative_advantage()). If it is a matrix it must be a
zero/one matrix with countries in the row names and products in the column names.
If rca is a tibble/data.frame it must contain the columns.}

\item{c1}{string to indicate the column that contains exporting countries in rca (e.g. "country")}

\item{p1}{string to indicate the column that contains exported products in rca (e.g. "product")}

\item{v1}{string to indicate the column that contains traded values in rca(e.g. "value")}

\item{proximity_products}{matrix or tibble/data.frame, if d is a tibble/data.frame it must contain the columns
from (character/factor), to (character/factor) and value (numeric), if it is a matrix it must be a
numeric matrix with products in row names and column names}

\item{p21}{string to indicate the column that contains exported products in proximity_products (e.g. "from")}

\item{p22}{string to indicate the column that contains exported products in proximity_products (e.g. "to")}

\item{v2}{string to indicate the column that contains proximity values in proximity_products (e.g. "value")}

\item{pci}{numeric vector with products in names (e.g. \code{product_complexity_index}}

\item{p3}{string to indicate the column that contains exported products in proximity_products (e.g. "product")}

\item{v3}{string to indicate the column that contains proximity values in proximity_products (e.g. "value")
from \code{complexity_measures()})}

\item{tbl_output}{when set to TRUE the output will be a tibble instead of a matrix (default set to FALSE)}
}
\description{
Countries position
}
\examples{
countries_position(
    rca = revealed_comparative_advantage_output,
    c1 = "country",
    p1 = "product",
    v1 = "value",
    proximity_products = proximity_matrices_output$products_proximity,
    p21 = "from",
    p22 = "to",
    v2 = "value",
    pci = complexity_measures_output$product_complexity_index,
    p3 = "product",
    v3 = "value"
)
}
\references{
For more information on proximity distance, complexity outlook, complexity outlook gain and its
applications see:

\insertRef{atlas2014}{economiccomplexity}
}
\keyword{functions}
