% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complexity_measures.R
\name{ec_complexity_measures}
\alias{ec_complexity_measures}
\title{Complexity Measures}
\usage{
ec_complexity_measures(rca, country = "country", product = "product",
  value = "value", method = "fitness", iterations = 20,
  extremality = 1, tbl = TRUE, atlas = FALSE)
}
\arguments{
\item{rca}{matrix or data.frame with RCA values}

\item{country}{column containing countries (applies only if d is a
data.frame)}

\item{product}{column containing products (applies only if d is a
data.frame)}

\item{value}{column containing traded values (applies only if d is a
data.frame)}

\item{method}{which method to use (by default is "fitness", it can also
be "reflections" or "eigenvalues")}

\item{iterations}{number of iterations to use in the fitness, reflections
and indirectly in eigenvalues method (by default is 20)}

\item{extremality}{numeric coefficient to use in the fitness method
(by default is 1)}

\item{tbl}{TRUE (default) returns a data.frame and FALSE returns a matrix}

\item{atlas}{remove the countries not ranked in The Atlas of Economic
Complexity (by default is FALSE)}
}
\description{
Given a \eqn{C\times P} matrix (C for "countries"
and P for "products") or an equivalent 3-columns data.frame with RCA
values as input, this function computes Country/Product Complexity Indices
and Diversity/Ubiquity metrics.
The equations used in this function are the generalized fitness
\eqn{F(\gamma)} and complexity \eqn{Q(\gamma)} defined by:
\deqn{\tilde{F}^{(n)}(\gamma) = \sum_{p} R_{cp}Q_{p}^{n-1}}
\deqn{\tilde{Q}^{(n)}(\gamma) =
\left[\sum_{c} R_{cp}(F_{c}^{n-1})^{-\gamma}\right]^{-1/\gamma}}
Altogether with equivalent formulations from other methods.
}
\examples{
ec_complexity_measures(ec_output_demo$rca_tbl)

}
\references{
For more information on complexity measures see:

\insertRef{measuringcomplexity2015}{economiccomplexity}

\insertRef{atlas2014}{economiccomplexity}

and the references therein.
}
\keyword{functions}
