% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/economic_complexity_measures.R
\name{economic_complexity_measures}
\alias{economic_complexity_measures}
\title{Economic Complexity Measures}
\usage{
economic_complexity_measures(d = NULL, c = "country", p = "product",
  v = "value", use_eigenvalues = FALSE, iterations = 20,
  tbl_output = FALSE)
}
\arguments{
\item{d}{matrix or tibble/data.frame in long format, if d is a tibble/data.frame it must contain the columns
country (character/factor), product (character/factor) and discrete RCA (integer), if it is a matrix it must be a
zero/one matrix with countries in the row names and products in the column names}

\item{c}{string to indicate the column that contains exporting countries (default set to "country" that is the
output of revealed_comparative_advantage())}

\item{p}{string to indicate the column that contains exported products (default set to "product" that is the
output of revealed_comparative_advantage())}

\item{v}{string to indicate the column that contains RCA values (default set to "value" that is the
output of revealed_comparative_advantage())}

\item{use_eigenvalues}{when set to TRUE the function will use eigenvalues as a limit case of the reflections
methods (default set to FALSE)}

\item{iterations}{number of iterations to use in the reflections method (default set to 20)}

\item{tbl_output}{when set to TRUE the output will be a tibble instead of a matrix (default set to FALSE)}
}
\description{
Economic Complexity Measures
}
\examples{
complexity_measures <- economic_complexity_measures(world_rca_2017)
}
\keyword{functions}
