% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/economiccomplexity-package.R
\docType{data}
\name{economiccomplexity_output}
\alias{economiccomplexity_output}
\title{Example Outputs of the Functions within the Package}
\format{
Lists of vectors, matrices, and graphs.
}
\usage{
data(economiccomplexity_output)
}
\description{
List of different computations obtained by using the functions within the package.
}
\examples{
data(economiccomplexity_output)
economiccomplexity_output$balassa_index
economiccomplexity_output$complexity_measures$complexity_index_country
}
\keyword{datasets}
