\name{economiccomplexity_output}
\alias{economiccomplexity_output}
\docType{data}
\title{
Example outputs of the functions within the package
}
\description{
List of different computations obtained by using the functions within the
package}
\usage{economiccomplexity_output}
\format{
  Lists of vectors, matrices and graphs.
}
\examples{
economiccomplexity_output$balassa_index
economiccomplexity_output$complexity_measures$complexity_index_country
}
\keyword{datasets}
