% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complexity_measures.R
\name{eigenvalues_method}
\alias{eigenvalues_method}
\title{Eigenvalues Method}
\usage{
eigenvalues_method(balassa_index, iterations)
}
\arguments{
\item{balassa_index}{(Type: dgCMatrix) the output from
\code{balassa_index()}) or an equivalent arrangement.}

\item{iterations}{(Type: numeric) the number of iterations to use.
By default this is set to \code{20}.}
}
\description{
Eigenvalues Method
}
\keyword{internal}
