% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networks.R
\name{networks}
\alias{networks}
\title{Networks}
\usage{
networks(pc, pp, c_cutoff = 0.35, p_cutoff = 0.55)
}
\arguments{
\item{pc}{matrix or tibble}

\item{pp}{matrix or tibble}

\item{c_cutoff}{numeric}

\item{p_cutoff}{numeric}
}
\description{
Networks
}
\examples{
rca <- rca(d = world_trade_1980, c = "reporter_iso",
    p = "product_code", x = "export_value_usd")

indices <- indices(rca, method = "reflections", maxiter = 20,
    output = "matrix")

m <- indices$m
kc0 <- indices$kc0
kp0 <- indices$kp0

proximity <- proximity(m, kc = kc0, kp = kp0, output = "matrix")
}
\keyword{functions}
