% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/productivity_levels.R
\name{productivity_levels}
\alias{productivity_levels}
\title{Productivity Levels}
\usage{
productivity_levels(data_exp, data_gdp, country = "country",
  product = "product", value = "value")
}
\arguments{
\item{data_exp}{(Type: data.frame, matrix or dgCMatrix) a dataset containing
countries, products and exported values.}

\item{data_gdp}{(Type: data.frame, matrix or dgCMatrix) a dataset
containing countries and per-capita GDP values.}

\item{country}{(Type: character) the column with the countries.
By default this is set to \code{"country"}. Used only if the input is a data.frame.}

\item{product}{(Type: character) the column with the products.
By default this is set to \code{"product"}. Used only if the input is a data.frame.}

\item{value}{(Type: character) the column with the metric for
country-product pairs.
By default this is set to \code{"value"}. Used only if the input is a data.frame.}
}
\value{
A list of two named numeric vectors.
}
\description{
\code{productivity_levels()} computes EXPY and PRODY measures.
}
\details{
The current implementation follows
\insertCite{exportmatters2005}{economiccomplexity} to obtain the
productivity and income levels associated to exports.
}
\examples{
data_exp <- galactic_federation

# Since I don't have a per-capita GDP dataset for the Galactic Federation,
# I'll create simulated data
set.seed(1810)
data_gdp <- setNames(
 rnorm(1:nrow(galactic_federation), 1000, 200),
 rownames(galactic_federation)
)

productivity_levels(data_exp, data_gdp)

}
\references{
For more information on prody and its applications see:

\insertRef{exportmatters2005}{economiccomplexity}
}
\keyword{functions}
