% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projections.R
\name{projections}
\alias{projections}
\title{Projections of a bipartite network}
\usage{
projections(
  proximity_country,
  proximity_product,
  avg_links = 4,
  tolerance = 0.05,
  compute = "both"
)
}
\arguments{
\item{proximity_country}{a data frame containing proximity
values for the elements of set X (e.g. proximity_country from \code{proximity()})}

\item{proximity_product}{a data frame containing proximity
values for the elements of set Y (e.g. proximity_product from \code{proximity()})}

\item{avg_links}{average number of connections for the projection of X
(default set to 4)}

\item{tolerance}{tolerance for proximity variation on each iteration until
obtaining the desired average number of connections (default set to 0.05)}

\item{compute}{which projection to compute. By default is "both" (both
projections) but it can also be "country" or "product".}
}
\description{
TBD
}
\details{
TBD
}
\examples{
projections(
  proximity_country = economiccomplexity_output$proximity$proximity_country,
  proximity_product = economiccomplexity_output$proximity$proximity_product
)
}
\references{
For more information see:

\insertRef{atlas2014}{economiccomplexity}

and the references therein.
}
\keyword{functions}
