% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proximity.R
\name{proximity}
\alias{proximity}
\title{Proximity}
\usage{
proximity(rca, diversity, ubiquity, compute = "both", tbl = TRUE,
  country_r = "country", product_r = "product", value_r = "value",
  country_d = "country", value_d = "value", product_u = "product",
  value_u = "value")
}
\arguments{
\item{rca}{matrix or data.frame with RCA values}

\item{diversity}{matrix or data.frame with diversity values}

\item{ubiquity}{matrix or data.frame with ubiquity values}

\item{compute}{"country", "product" or "both" (default) matrices}

\item{tbl}{TRUE (default) returns a data.frame and FALSE returns a matrix}

\item{country_r}{column containing countries (applies only if d is a
data.frame)}

\item{product_r}{column containing products (applies only if d is a
data.frame)}

\item{value_r}{column containing traded values (applies only if d is a
data.frame)}

\item{country_d}{column containing countries (applies only if d is a
data.frame)}

\item{value_d}{column containing values (applies only if d is a
data.frame)}

\item{product_u}{column containing products (applies only if d is a
data.frame)}

\item{value_u}{column containing values (applies only if d is a
data.frame)}
}
\value{
A list with two data frames or matrices.
}
\description{
\code{proximity} computes proximity

Given a \eqn{C\times P} matrix (C for "countries"
and P for "products") with RCA values, a C lenght vector with diversity
values and a P length vector with ubiquity values or equivalent data frames,
this function implements the equations:
\deqn{\text{(Product proximity)}\: \phi_{pp'} = \frac{\sum_c R_{cp}
R_{cp'}}{\max(k_{p}^{(0)}, k_{p'}^{(0)})}}
\deqn{\text{(Country proximity)}\:\lambda_{cc'} = \frac{\sum_p R_{c,p}
R_{c,p'}}{\max(k_{c}^{(0)}, k_{c'}^{(0)})}}
}
\examples{
proximity(
  ec_output_demo$rca,
  ec_output_demo$complexity$diversity,
  ec_output_demo$complexity$ubiquity
)

}
\references{
For more information on proximity and its applications see:

\insertRef{atlas2014}{economiccomplexity}

and the references therein.
}
\seealso{
\code{\link[economiccomplexity]{rca}},
\code{\link[economiccomplexity]{complexity}}
}
\keyword{functions}
