% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proximity_matrices.R
\name{proximity_matrices}
\alias{proximity_matrices}
\title{Proximity Matrices}
\usage{
proximity_matrices(d = NULL, c = "country", p = "product",
  v = "value", diversity = NULL, ubiquity = NULL,
  diversity_c = "country", diversity_v = "value",
  ubiquity_p = "product", ubiquity_v = "value", tbl_output = FALSE)
}
\arguments{
\item{d}{matrix or tibble/data.frame in long format, if d is a tibble/data.frame it must contain the columns
country (character/factor), product (character/factor) and discrete RCA (integer), if it is a matrix it must be a
zero/one matrix with countries in the row names and products in the column names}

\item{c}{string to indicate the column that contains exporting countries (default set to "country" that is the
output of revealed_comparative_advantage())}

\item{p}{string to indicate the column that contains exported products (default set to "product" that is the
output of revealed_comparative_advantage())}

\item{v}{string to indicate the column that contains RCA values (default set to "value" that is the
output of revealed_comparative_advantage())}

\item{diversity}{vector or tibble/data.frame containing diversity measures}

\item{ubiquity}{vector or tibble/data.frame containing ubiquity measures}

\item{diversity_c}{string to indicate the column that contains diversity countries (default set to "country" that is the
output of economic_complexity_measures())}

\item{diversity_v}{string to indicate the column that contains diversity values (default set to "value" that is the
output of economic_complexity_measures())}

\item{ubiquity_p}{string to indicate the column that contains diversity products (default set to "product" that is the
output of economic_complexity_measures())}

\item{ubiquity_v}{string to indicate the column that contains ubiquity values (default set to "value" that is the
output of economic_complexity_measures())}

\item{tbl_output}{when set to TRUE the output will be a tibble instead of a matrix (default set to FALSE)}
}
\description{
Proximity Matrices
}
\examples{
proximity <- proximity_matrices(d = world_rca_2017,
    diversity = complexity_measures_2017$diversity,
    ubiquity = complexity_measures_2017$ubiquity)
}
\keyword{functions}
