% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rca.R
\name{rca}
\alias{rca}
\title{Revealed Comparative Advantage (RCA)}
\usage{
rca(data, country = "country", product = "product", value = "value",
  discrete = TRUE, cutoff = 1, tbl = TRUE)
}
\arguments{
\item{data}{matrix or data.frame with traded values}

\item{country}{column containing countries (applies only if d is a
data.frame)}

\item{product}{column containing products (applies only if d is a
data.frame)}

\item{value}{column containing traded values (applies only if d is a
data.frame)}

\item{discrete}{convert to one all the values above a cutoff and zero
otherwise (by default is TRUE)}

\item{cutoff}{all the values below the specified will be converted to zero
(by default is 1)}

\item{tbl}{TRUE (default) returns a data.frame and FALSE returns a matrix}
}
\value{
A list with two data frames or two matrices.
}
\description{
\code{rca} computes Revealed Comparative Advantage
}
\details{
Given a \eqn{C\times P} matrix (C for "countries"
and P for "products") or an equivalent 3-columns data frame with exported
values (X) as input, this function implements the equation:
\deqn{RCA_{cp} = \frac{X_{cp}}{\sum_c X_{cp}} /
\frac{\sum_p X_{cp}}{\sum_{c}\sum_{p} X_{cp}}}
}
\examples{
rca(ec_trade_1962)

}
\references{
For more information on revealed comparative advantage and its uses see:

\insertRef{atlas2014}{economiccomplexity}
}
\keyword{functions}
