% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rca.R
\name{rca}
\alias{rca}
\title{Computes RCA (Revealed Comparative Advantage)}
\usage{
rca(data = data, c = "country", p = "product", value = "export_val")
}
\arguments{
\item{data}{a tibble (or data frame) in long format}

\item{c}{name of column that contains countries (by default is "country")}

\item{p}{name of column that contains products (by default is "product")}

\item{value}{name of column that contains trade values (by default is "export_val")}
}
\description{
Computes RCA (Revealed Comparative Advantage)
}
\examples{
# Demo dataset 'Fantasy World'
rca(fantasy_world_long)
}
\keyword{functions}
