% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rca.R
\name{rca}
\alias{rca}
\title{Computes RCA (Revealed Comparative Advantage)}
\usage{
rca(d = NULL, c = NULL, p = NULL, v = NULL)
}
\arguments{
\item{d}{a tibble (or data frame) in long format}

\item{c}{column that contains exporting countries (e.g. "reporter_iso")}

\item{p}{column that contains products (e.g. "product_code)}

\item{v}{column that contains traded values (e.g. "trade_value_usd")}
}
\description{
Computes RCA (Revealed Comparative Advantage)
}
\examples{
# TBD
}
\keyword{functions}
