% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revealed_comparative_advantage.R
\name{revealed_comparative_advantage}
\alias{revealed_comparative_advantage}
\title{Revealed Comparative Advantage (RCA)}
\usage{
revealed_comparative_advantage(d = NULL, c = NULL, p = NULL,
  v = NULL, cutoff = 1, discrete = TRUE, tbl_output = FALSE)
}
\arguments{
\item{d}{tibble/data.frame in long format, it must contain the columns country (character/factor),
product (character/factor) and export value (numeric)}

\item{c}{string to indicate the column that contains exporting countries (e.g. "reporter_iso")}

\item{p}{string to indicate the column that contains exported products (e.g. "product_code")}

\item{v}{string to indicate the column that contains traded values (e.g. "trade_value_usd")}

\item{cutoff}{when set to TRUE all the values lower than the specified cutoff will be
converted to zero and to one in other case, numeric (default set to 1)}

\item{discrete}{when set to TRUE it will convert all the Revealed Comparative Advantage values
to zero or one based on the cutoff value (default set to TRUE)}

\item{tbl_output}{when set to TRUE the output will be a tibble instead of a matrix (default set to FALSE)}
}
\description{
Revealed Comparative Advantage (RCA)
}
\examples{
rca <- revealed_comparative_advantage(d = world_trade_2017, c = "reporter_iso",
    p = "product_code", v = "export_value_usd")
}
\references{
For more information on revealed comparative advantage and its uses see:

\insertRef{atlas2014}{economiccomplexity}
}
\keyword{functions}
