\name{revealed_comparative_advantage_output}
\alias{revealed_comparative_advantage_output}
\docType{data}
\title{
Demo dataset to show the output of revealed_comparative_advantage().
}
\description{
TBD
}
\usage{data("revealed_comparative_advantage_output")}
\format{
  A data frame with 176971 observations on the following 3 variables.
  \describe{
    \item{\code{country}}{a character vector}
    \item{\code{product}}{a character vector}
    \item{\code{value}}{a numeric vector}
  }
}
\source{
Open Trade Statistics (obtained by using the tradestatistics package)
}
\examples{
revealed_comparative_advantage_output
}
\keyword{datasets}
