% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/economiccomplexity-package.R
\docType{data}
\name{world_gdp_avg_1998_to_2000}
\alias{world_gdp_avg_1998_to_2000}
\title{World Trade Per-Capita GDP for the Period 1998-2000}
\format{
A data frame with 240 rows and 2 columns.
}
\usage{
data(world_gdp_avg_1998_to_2000)
}
\description{
A data frame that summarizes the per-capita GDP of different countries. The unit is year 2000 USD.
}
\examples{
data(world_gdp_avg_1998_to_2000)
head(world_gdp_avg_1998_to_2000)
}
\keyword{datasets}
