\name{world_gdp_avg_1998_to_2000}
\alias{world_gdp_avg_1998_to_2000}
\docType{data}
\title{
World Trade Per-Capita GDP for the Period 1998-2000
}
\description{
A data frame that summarises the per-capita GDP of different countries.
The unit is year 2000 USD.
}
\usage{world_gdp_avg_1998_to_2000}
\format{
  A data frame with 240 rows and 2 columns.
}
\examples{
world_gdp_avg_1998_to_2000
}
\keyword{datasets}
