\name{world_rca_2017}
\alias{world_rca_2017}
\docType{data}
\title{
Demo dataset to show the output of revealed_comparative_advantage().
}
\description{
TBD
}
\usage{data("world_rca_2017")}
\format{
  A data frame with 176971 observations on the following 3 variables.
  \describe{
    \item{\code{country}}{a character vector}
    \item{\code{product}}{a character vector}
    \item{\code{value}}{a numeric vector}
  }
}
\source{
Open Trade Statistics (obtained by using the tradestatistics package)
}
\examples{
world_rca_2017
}
\keyword{datasets}
