\name{world_trade_2017}
\alias{world_trade_2017}
\docType{data}
\title{
tbd
}
\description{
tbd
}
\usage{data("world_trade_2017")}
\format{
  A data frame with 165977 observations on the following 3 variables.
  \describe{
    \item{\code{reporter_iso}}{a character vector}
    \item{\code{product_code}}{a character vector}
    \item{\code{export_value_usd}}{a numeric vector}
  }
}
\details{
tbd
}
\source{
tbd
}
\references{
tbd
}
\examples{
data(world_trade_2017)
## maybe str(world_trade_2017) ; plot(world_trade_2017) ...
}
\keyword{datasets}
