\name{world_trade_avg_1998_to_2000}
\alias{world_trade_avg_1998_to_2000}
\docType{data}
\title{
World Trade Averages for the Period 1998-2000
}
\description{
A matrix that summarises all the products that
different countries exported to the rest of the world. This data uses the SITC
revision 2 classification with four digits product codes. The unit is year 2000
USD.
}
\usage{world_trade_avg_1998_to_2000}
\format{
  A sparse matrix with 226 rows and 785 columns.
}
\examples{
world_trade_avg_1998_to_2000
}
\keyword{datasets}
